pyproject.toml
[project]
name = "blackbird"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.123.0",
    "ipykernel>=7.1.0",
    "pika>=1.3.2",
    "psycopg2-binary>=2.9.11",
    "pybit>=5.13.0",
    "pydantic>=2.12.5",
    "sortedcontainers>=2.4.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
]
